Move 'evidence-item' folder to 'components'
