+++ src/ggrc-client/js/components/evidence-item/templates/evidence-item.stache
